add api and players do_turn functions

+++ classes/ship.py
@@ -3,7 +3,7 @@ import utils
 
 
 class Ship:
-    def __init__(self, player_id: int, ship_speed: float, location: tuple):
+    def __init__(self, player_id: int, ship_speed: int, location: tuple):
         self.player_id = player_id
         self.ship_speed = ship_speed
         self.location = location
